+++ marshmallow/exceptions.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-'''Exception classes for marshmallow-related errors.'''
+"""Exception classes for marshmallow-related errors."""
 
 class MarshmallowError(Exception):
     '''Base class for all marshmallow-related errors.'''
@@ -14,6 +14,7 @@ class MarshallingError(MarshmallowError):
     """
 
     def __init__(self, underlying_exception):
+        self.underlying_exception = underlying_exception
         # just put the contextual representation of the error to hint on what
         # went wrong without exposing internals
         super(MarshallingError, self).__init__(str(underlying_exception))